+++ src/components/BatchAllocations/UploadBatchAllocations/UploadBatchAllocations.test.js
@@ -156,6 +156,27 @@ describe('UploadBatchAllocations', () => {
     expect(defaultProps.history.push).toHaveBeenCalled();
   });
 
+  it('should handle batch allocation error', async () => {
+    handle.mockRejectedValueOnce();
+    useBatchAllocation.mockReturnValue({ budgetsFunds: [fyFinanceData[0]] });
+
+    renderComponent();
+
+    await waitFor(() => {
+      expect(screen.getByText('test.csv')).toBeInTheDocument();
+    });
+
+    await userEvent.type(screen.getAllByLabelText('ui-finance.transaction.allocation.batch.columns.budgetAllocationChange')[0], '42');
+    await userEvent.click(screen.getByRole('button', { name: 'ui-finance.allocation.batch.form.footer.recalculate' }));
+    await userEvent.click(screen.getByRole('button', { name: 'stripes-components.saveAndClose' }));
+
+    expect(handle).toHaveBeenCalled();
+    expect(showCalloutMock).toHaveBeenCalledWith({
+      messageId: 'ui-finance.actions.allocations.batch.error',
+      type: 'error',
+    });
+  });
+
   it('should call recalculate on recalculate button click', async () => {
     renderComponent();